feature: added sending and getting values by Kafka

+++ src/main/java/attachment/service/attachmentservice/controller/AttachmentController.java
@@ -16,6 +16,7 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 
 @RestController
@@ -24,6 +25,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class AttachmentController {
     private final TokenValidation tokenValidation;
     private final AttachmentService attachmentService;
+    private final RestTemplate restTemplate;
 
     @PostMapping
     public ResponseEntity<String> uploadFile(@RequestParam("taskId") Long taskId,
@@ -49,28 +51,52 @@ public class AttachmentController {
     }
 
     @GetMapping
-    public ResponseEntity<String> downloadFile(@RequestParam Long taskId) {
-        try {
-            String link = attachmentService.downloadFromDropBox(taskId);
-            return ResponseEntity.ok(link);
-        } catch (DbxException e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body("Failed to download the file. Please try again later.");
-        } catch (EntityNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body("Attachment not found for the given task ID.");
+    public ResponseEntity<String> downloadFile(@RequestParam Long taskId,
+                                               @RequestHeader("Authorization") String token) {
+        tokenValidation.sendTokenToValidate(token);
+
+        boolean isValid = tokenValidation.waitForTokenValidation(token);
+
+        if (isValid) {
+            try {
+                String link = attachmentService.downloadFromDropBox(taskId);
+                return ResponseEntity.ok(link);
+            } catch (DbxException e) {
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                        .body("Failed to download the file. Please try again later.");
+            } catch (EntityNotFoundException e) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body("Attachment not found for the given task ID.");
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteById(@PathVariable Long id) {
-        try {
-            attachmentService.deleteAttachment(id);
-            return ResponseEntity.status(HttpStatus.ACCEPTED)
-                    .body("Attachment dy id: " + id + "was deleted");
-        } catch (EntityNotFoundException | DbxException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body("Attachment not found for the given task ID: " + id);
+    public ResponseEntity<String> deleteById(@PathVariable Long id,
+                                             @RequestHeader("Authorization") String token) {
+        tokenValidation.sendTokenToValidate(token);
+
+        boolean isValid = tokenValidation.waitForTokenValidation(token);
+        if (isValid) {
+            try {
+                attachmentService.deleteAttachment(id);
+                return ResponseEntity.status(HttpStatus.ACCEPTED)
+                        .body("Attachment dy id: " + id + "was deleted");
+            } catch (EntityNotFoundException | DbxException e) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body("Attachment not found for the given task ID: " + id);
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token");
         }
+    }
+
+    private boolean validateToken(String token) {
+        String validateEndPoint = "http://localhost:8080/api/auth/validate";
+        ResponseEntity<String> response = restTemplate.postForEntity(validateEndPoint, token, String.class);
+
+        return response.getStatusCode().equals(HttpStatus.OK);
     }
 }
